Removed config, updated Behat command to extend DuskCommand, updated tests
