--- config/behat-dusk.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'dusk_base_url' => env('DUSK_BASE_URL') ?? 'http://localhost'
-];
